+++ src/append/file/mod.rs
@@ -31,7 +31,7 @@
 //!
 //! logforth::builder()
 //!     .dispatch(|d| d.filter(LevelFilter::Trace).append(rolling_file))
-//!     .apply();
+//!     .setup_log_crate();
 //!
 //! log::info!("This log will be written to a rolling file.");
 //! ```