+++ tests/recursive_logging.rs
@@ -49,7 +49,7 @@ fn test_meta_logging_in_format_works() {
         .dispatch(|d| d.append(stdout))
         .dispatch(|d| d.append(stderr))
         .dispatch(|d| d.append(rolling))
-        .apply();
+        .setup_log_crate();
 
     struct Thing<'a>(&'a str);