+++ CHANGELOG.md
@@ -39,6 +39,7 @@ All notable changes to this project will be documented in this file.
 * `EnvFilter` is now self-hosted. Some methods may be changed, but the general user experience should retain:
   * `EnvFilter`'s constructors (`from_env`, etc.) are moved to `EnvFilterBuilder`.
 * Upgrade to opentelemetry 0.31.0.
+* `LoggerBuilder`'s `try_apply`/`apply` methods are renamed to `try_setup_log_crate`/`setup_log_crate`.
 * Internal log structs are migrated from `log` crate to self-hosted types. This should not affect most users, but if you are customizing appender, layout, filter, and diagnostic, you should replace `log::Record`, `log::Metadata`, or `log::Level`, with `logforth::Record`, `logforth::Metadata`, or `logforth::Level`.
 
 ### Notable changes